requirements.txt
python-dotenv
openai
pdf2image

asyncpg
FlagEmbedding==1.2.10
transformers==4.40.2
huggingface_hub==0.23.0
torch>=2.1.0